pyproject.toml
[project]
name = "slack-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2025.4.26",
    "mcp[cli]>=1.9.0",
    "pyyaml>=6.0.2",
    "slack-sdk>=3.35.0",
]
